refactor(extractor): simplify binary search
This reimplementation of the binary search used
to find indices of GitHub API items was prompted
by the recognition that, if an item was missing
from the paginated list, e.g. an issue didn't
appear in it because it was removed or something,
and was supposed to be at the last index of a
page, the former implementation would return
the wrong page to the binary search that looked
for the desired item in the found page, which
was assumed correct. This means that, in some
rare occasion that we haven't witnessed yet, it
would return an incorrect index.

This implementation avoids finding the "correct"
page and passing it and, instead, "denests" the
item numbers from the paginated list; instead of
finding the right page (top list), then finding
the item in it (inner list), this way simply
treats the paginated list as a list with one
layer (where, in reality, it is a list of lists).
The intention is to avoid issues where the wrong
page is selected and passed to the binary search
that searches it.

Concerning sloc, it is significantly smaller.
Concerning operational speed, it is
significantly slower according to
time.perf_counter(): where the former
implementation took 4~ seconds, this one takes
6~.