refactor (file io): rm fast failure for json FileNotFoundError
read_jsonfile_into_dict() failed fast using
sys.exit() if an input json was not found to
read from at the path it was given. This meant
that, anytime the extractor tried to do a
recursive dict merge and write but couldn't find
the output file, the whole program died and all
collected data was lost. Instead of failing fast,
we return an empty dictionary and check the
contents on the other side, if needed.

For example, this method is used when reading
the input configuration (which is in json). If
that is empty, we don't need to fail fast because
our schema validator will check that. This
method is also used to read the
already-collected data out of the output file
before newly collected data is merged into it.
If that file doesn't exist yet, we don't need to
terminate over it. We can just create the file
and keep working