perf: parallelize range indices finding
Instead of finding the indices of the start and
end range values independently, we can
parallelize the binary search for them.

time.perf_counter() reports that this process
takes about 5.5 to 6 seconds. This halves the
time for the new binary search reported in
0649392 and reduces runtime by 25% (about 8-9
seconds to 5.5-6 seconds) from the original
implementation.

- readd printing for the index finding process
- add a docstring for the method.
- add a try-except-else to prevent program
  termination if we are rate limited before or
  while getting range indices