feat, refactor(sleep): poll API to protect from unnecessary execution
For a long time, I didn't understand why the
sleep method for the extractor would print its
logging repeatedly. I realized that it was
returning when the countdown finished and the
except block that ran it would simply run again
because the countdown until rate limiting ends
is not absolutely precise.

This fix, which may be temporary if I figure
out a smarter way to handle the countdown,
checks if calls can be made again, sleeps if they
cant, then checks again. It appears like Github
doesn't send an event for when rate limiting
ends and, as a consequence, I poll instead of
listen. If they did send an event, the function
to get the remaining rate limit might not be
necessary